texture-app/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
black==23.11.0
isort==5.12.0
requests==2.32.3